requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
supabase==2.4.0
pydantic==2.6.3
python-multipart==0.0.9
python-dotenv==1.0.1
